Embed inclua_lua.hpp in generated wrapper, Lua userdata
UPDATE: no need for .hpp: just embed it =]

FIX: Lua light userdata share the same metatable (didn't know that...),
so it all goes BOOM. Now using only full userdata, and so every pointer
now is pushed as a full userdata. There is a need to distinguish which
ones are pointers, and which are not, so using uservalues now as well.
TODO: make '__eq' metamethod, so that the same pointer returns `true` on `==`