-I indicates where to find headers, cmake Use module
NEW: try to parse headers not only in current directory, but also in
'-I' indicated directories. This eases the use of the same inclua script
for different platforms =]

NEW: cmake UseInclua module, so it's easy to use inclua with cmake.
More build systems will be added later, including a cmake Find module