Lua: added void* push/check, fixed delete char*, initialize out pointers
NEW: `void *` template specialization for `inclua_push` and
`inclua_check`. Pushes light userdata, and on check don't worry about
metatables, but still check for full userdata (if it is a pointer or
not)

FIX: removed the `size` argument from `inclua_delete_array` for strings,
now it calls the right overload.

NEW: initializing output parameters that are pointers to `NULL`. Some
functions expect this guarantee, so always do it.